+++ src/views/domain-page/__fixtures__/domain-page-query-params.ts
@@ -27,4 +27,6 @@ export const mockDomainPageQueryParamsValues = {
   queryArchival: '',
   clusterAttributeScope: undefined,
   clusterAttributeValue: undefined,
+  schedulesSearch: '',
+  schedulesStatus: undefined,
 } as const satisfies PageQueryParamValues<typeof domainPageQueryParamsConfig>;
