+++ src/views/domain-schedules/config/domain-schedules-filters.config.ts
@@ -1,13 +1,18 @@
+import { createElement } from 'react';
+
+import ListFilter from '@/components/list-filter/list-filter';
 import { type PageFilterConfig } from '@/components/page-filters/page-filters.types';
+import type domainPageQueryParamsConfig from '@/views/domain-page/config/domain-page-query-params.config';
 
-import DomainSchedulesFiltersStatus from '../domain-schedules-filters-status/domain-schedules-filters-status';
-import { type DomainSchedulesFiltersStatusValue } from '../domain-schedules-filters-status/domain-schedules-filters-status.types';
-
-import type domainSchedulesQueryParamsConfig from './domain-schedules-query-params.config';
+import { DOMAIN_SCHEDULES_STATUS_LABELS_MAP } from '../domain-schedules.constants';
+import type {
+  DomainSchedulesFiltersStatusValue,
+  DomainSchedulesStatus,
+} from '../domain-schedules.types';
 
 const domainSchedulesFiltersConfig: [
   PageFilterConfig<
-    typeof domainSchedulesQueryParamsConfig,
+    typeof domainPageQueryParamsConfig,
     DomainSchedulesFiltersStatusValue
   >,
 ] = [
@@ -15,7 +20,14 @@ const domainSchedulesFiltersConfig: [
     id: 'status',
     getValue: (v) => ({ schedulesStatus: v.schedulesStatus }),
     formatValue: (v) => v,
-    component: DomainSchedulesFiltersStatus,
+    component: ({ value, setValue }) =>
+      createElement(ListFilter<DomainSchedulesStatus>, {
+        label: 'Status',
+        placeholder: 'All statuses',
+        value: value.schedulesStatus,
+        onChangeValue: (v) => setValue({ schedulesStatus: v }),
+        labelMap: DOMAIN_SCHEDULES_STATUS_LABELS_MAP,
+      }),
   },
 ] as const;
