--- src/views/domain-schedules/domain-schedules-filters-status/domain-schedules-filters-status.types.ts
@@ -1,5 +1,0 @@
-export type DomainSchedulesStatus = 'RUNNING' | 'PAUSED';
-
-export type DomainSchedulesFiltersStatusValue = {
-  schedulesStatus: DomainSchedulesStatus | undefined;
-};
