+++ src/views/domain-schedules/domain-schedules.types.ts
@@ -1,3 +1,9 @@
 import { type DomainPageTabContentProps } from '@/views/domain-page/domain-page-content/domain-page-content.types';
 
+export type DomainSchedulesStatus = 'RUNNING' | 'PAUSED';
+
+export type DomainSchedulesFiltersStatusValue = {
+  schedulesStatus: DomainSchedulesStatus | undefined;
+};
+
 export type Props = DomainPageTabContentProps;
