+++ src/views/domain-schedules/helpers/is-domain-schedules-status.ts
@@ -1,5 +1,5 @@
-import { DOMAIN_SCHEDULES_STATUS_LABELS_MAP } from '../domain-schedules-filters-status/domain-schedules-filters-status.constants';
-import { type DomainSchedulesStatus } from '../domain-schedules-filters-status/domain-schedules-filters-status.types';
+import { DOMAIN_SCHEDULES_STATUS_LABELS_MAP } from '../domain-schedules.constants';
+import { type DomainSchedulesStatus } from '../domain-schedules.types';
 
 export function isDomainSchedulesStatus(
   value: string
